refactor: separated UserManagement into User Table, User Row, SearchFilter
